new file structure
New file structure, less knowlege in each file and everything is in the same includes.php. The config is in a folder config (we can also think about having some config for the dev and some other in production). ./src regroups all files to manage something.
