+++ includes.php
@@ -1,0 +1,3 @@
+<?php
+    require_once('config/params.php');
+    require_once('src/manage_keys.php');
