+++ src/.htaccess
@@ -1,0 +1,1 @@
+Options -Indexes
